requirements.txt
pandas
numpy
matplotlib
scikit-learn
pytorch
torchvision

